+++ src/main/java/com/trepicoder/adventurecraft/Adventurecraft.java
@@ -1,5 +1,9 @@
 package com.trepicoder.adventurecraft;
 
+import com.trepicoder.adventurecraft.events.BlockEvents;
+import net.minecraft.advancements.Advancement;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -10,6 +14,8 @@ import net.minecraftforge.fml.event.server.FMLServerStoppingEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.Collection;
+
 @Mod("adventurecraft")
 public class Adventurecraft {
     public static final String MODID = "adventurecraft";
@@ -18,6 +24,7 @@ public class Adventurecraft {
     public Adventurecraft() {
         // Register the mod class to handle event subscription
         MinecraftForge.EVENT_BUS.register(this);
+        MinecraftForge.EVENT_BUS.register(new BlockEvents(LOGGER));
     }
 
     @SubscribeEvent
@@ -41,7 +48,14 @@ public class Adventurecraft {
      * Handle server startup events (includes single player "server")
      */
     public void onServerStarting(FMLServerStartingEvent event) {
+        MinecraftServer server = event.getServer();
 
+        LOGGER.info("Adventurecraft is deregistering its own datapack so that it overrides every other datapack");
+        server.getCommandManager().handleCommand(server.getCommandSource().withPermissionLevel(2), "datapack disable \"mod:adventurecraft\"");
+
+
+        LOGGER.info("Clearing default advancements");
+        server.getAdvancementManager().getAllAdvancements().removeIf(advancement -> advancement.getId().getNamespace().contains("minecraft"));
     }
 
     @SubscribeEvent