+++ src/main/java/com/trepicoder/adventurecraft/blocks/ACWoodLogBlock.java
@@ -1,6 +1,5 @@
 package com.trepicoder.adventurecraft.blocks;
 
-import net.minecraft.block.Blocks;
 import net.minecraft.block.RotatedPillarBlock;
 
 public class ACWoodLogBlock extends RotatedPillarBlock {