+++ src/main/java/com/trepicoder/adventurecraft/registries/ACBlocks.java
@@ -4,6 +4,12 @@ import net.minecraft.block.Block;
 
 public class ACBlocks {
     //region Wood
+    //region Logs
     public static Block sugarmaple_log;
+    //endregion
+
+    //region Planks
+    public static Block sugarmaple_planks;
+    //endregion
     //endregion
 }