+++ src/main/java/com/trepicoder/adventurecraft/tools/ACWoodenTreetap.java
@@ -1,0 +1,23 @@
+package com.trepicoder.adventurecraft.tools;
+
+import net.minecraft.client.util.ITooltipFlag;
+import net.minecraft.item.AxeItem;
+import net.minecraft.item.IItemTier;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.world.World;
+
+import javax.annotation.Nullable;
+import java.util.List;
+
+public class ACWoodenTreetap extends AxeItem {
+    public ACWoodenTreetap(IItemTier tier, float attackDamageIn, float attackSpeedIn, Properties builder) {
+        super(tier, attackDamageIn, attackSpeedIn, builder);
+    }
+
+    @Override
+    public void addInformation(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+        tooltip.add(new StringTextComponent("Sappy Wappy Sap Tap"));
+    }
+}